sleepUntilTWIMatch Plan (3)
Decided not to include sleepUntilTWIMatch in LowPower.h and LowPower.cpp because it is not easy to get Wire.h from the libraries folder. Instead, that function will be put in a .ino file for a slave microcontroller that will go to sleep. It will be awaken by a master microcontroller.