MALS-110 Finished
Finished the test case for MALS-110. Use the DA board as the slave and the XPlained Mini board as the master. The two .ino files will run automatically. To validate this for MALS 130, use a power debugger to draw the graph and included it in a README.md file. Also include any additional comments and changes of the scripts to the README file.